Create syncer types file
